+++ includes/object-cache.php
@@ -2900,26 +2900,36 @@ LUA;
             die();
         }
 
+        $verbose = wp_installing()
+            || defined( 'WP_ADMIN' )
+            || ( defined( 'WP_DEBUG' ) && WP_DEBUG );
+
         $message = '<h1>' . __( 'Error establishing a Redis connection' ) . "</h1>\n";
 
-        if ( wp_installing() || defined( 'WP_ADMIN' ) || ( defined( 'WP_DEBUG' ) && WP_DEBUG ) ) {
+        if ( $verbose ) {
             $message .= "<p><code>" . $exception->getMessage() . "</code></p>\n";
 
             $message .= '<p>' . sprintf(
-                __( 'This means that the connection information in your %1$s file are incorrect or that the Redis server is unreachable.' ),
+                __( 'WordPress is unable to establish a connection to Redis. This means that the connection information in your %1$s file are incorrect, or that the Redis server is not reachable.' ),
                 '<code>wp-config.php</code>'
             ) . "</p>\n";
 
             $message .= "<ul>\n";
-            $message .= '<li>' . __( 'Are you sure you have the correct Redis host and port?' ) . "</li>\n";
-            $message .= '<li>' . __( 'Are you sure Redis server is running?' ) . "</li>\n";
+            $message .= '<li>' . __( 'Is the correct Redis host and port set?' ) . "</li>\n";
+            $message .= '<li>' . __( 'Is the Redis server running?' ) . "</li>\n";
             $message .= "</ul>\n";
 
             $message .= '<p>' . sprintf(
-                __( 'If you need help, please read the <a href="%s">installation instructions</a>.' ),
+                __( 'If you need help, please read the <a href="%1$s">installation instructions</a>.' ),
                 __( 'https://github.com/rhubarbgroup/redis-cache/blob/develop/INSTALL.md' )
             ) . "</p>\n";
         }
+
+        $message .= '<p>' . sprintf(
+            __( 'To disable Redis, delete the %1$s file in the %2$s directory.' ),
+            '<code>object-cache.php</code>',
+            '<code>/wp-content/</code>',
+        ) . "</p>\n";
 
         wp_die( $message );
     }